Split multi-type core files
