+++ src/WinPrint.Core/Abstractions/GraphicsFontUnit.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public enum GraphicsFontUnit
+{
+    Point,
+    Pixel
+}
