+++ src/WinPrint.Core/Abstractions/GraphicsStringFormat.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public sealed class GraphicsStringFormat
+{
+    public GraphicsStringFormatFlags FormatFlags { get; set; } = GraphicsStringFormatFlags.None;
+    public GraphicsTextAlignment Alignment { get; set; } = GraphicsTextAlignment.Near;
+    public GraphicsTextAlignment LineAlignment { get; set; } = GraphicsTextAlignment.Near;
+    public GraphicsStringTrimming Trimming { get; set; } = GraphicsStringTrimming.None;
+}
