+++ src/WinPrint.Core/Abstractions/GraphicsStringTrimming.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public enum GraphicsStringTrimming
+{
+    None
+}
