+++ src/WinPrint.Core/Abstractions/GraphicsTextAlignment.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public enum GraphicsTextAlignment
+{
+    Near,
+    Center,
+    Far
+}
