+++ src/WinPrint.Core/Abstractions/GraphicsTextRenderingMode.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public enum GraphicsTextRenderingMode
+{
+    Default,
+    ClearTypeGridFit
+}
