+++ src/WinPrint.Core/Abstractions/IGraphicsBrush.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public interface IGraphicsBrush : IGraphicsResource
+{
+}
