+++ src/WinPrint.Core/Abstractions/IGraphicsFont.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public interface IGraphicsFont : IGraphicsResource
+{
+}
