+++ src/WinPrint.Core/Abstractions/IGraphicsPen.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public interface IGraphicsPen : IGraphicsResource
+{
+}
