+++ src/WinPrint.Core/Abstractions/IGraphicsResource.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public interface IGraphicsResource : IDisposable
+{
+}
