+++ src/WinPrint.Core/Abstractions/IGraphicsState.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace WinPrint.Core.Abstractions;
+
+public interface IGraphicsState
+{
+}
