+++ src/WinPrint.Core/ContentTypeEngines/TextMateCte.cs
@@ -20,21 +20,6 @@ using TextMateFontStyle = TextMateSharp.Themes.FontStyle;
 
 namespace WinPrint.Core.ContentTypeEngines;
 
-internal sealed class TextMateWrappedRun
-{
-    public int Start { get; init; }
-    public int Length { get; init; }
-    public Color Foreground { get; init; } = Color.Black;
-    public TextMateFontStyle FontStyle { get; init; } = TextMateFontStyle.None;
-}
-
-internal sealed class TextMateWrappedLine
-{
-    public int NonWrappedLineNumber { get; init; }
-    public string Text { get; init; } = string.Empty;
-    public List<TextMateWrappedRun> Runs { get; } = [];
-}
-
 /// <summary>
 ///     Syntax-highlighting content engine backed by TextMateSharp bundled grammars.
 /// </summary>
