+++ src/WinPrint.Core/Models/Footer.cs
@@ -1,0 +1,9 @@
+using System.Linq;
+using System.Text.Json.Serialization;
+using System.Text.RegularExpressions;
+
+namespace WinPrint.Core.Models;
+
+public class Footer : HeaderFooter
+{
+}
